move overlay trigger to sharinglabel component
