+++ src/options/pages/blueprints/gridView/GridCard.module.scss
@@ -51,7 +51,6 @@
 }
 
 .sharing {
-  cursor: help;
   min-width: 0;
 }
