+++ src/options/pages/blueprints/listView/ListItem.module.scss
@@ -49,6 +49,8 @@
 }
 
 .sharing {
+  display: flex;
+  flex-direction: column;
   flex-grow: 1;
   text-align: right;
   color: darkgray;
